Extract full v1 driver to common
